+++ functions/planner.h
@@ -3,18 +3,18 @@
 
 #include <stdbool.h>
 
-void planners_2D(double* matriz, 
+void planners_2D(float* matriz, 
                  int* size_map, 
-                 double* objective_points, 
+                 float* objective_points, 
                  int size_objective[2], 
-                 double* start_points, int size_start[2], 
+                 float* start_points, int size_start[2], 
                  int planner_type, int escalado_vectores);
 
-void planners_3D(double* matriz, 
+void planners_3D(float* matriz, 
                   int* size_map, 
-                  double* objective_points, 
+                  float* objective_points, 
                   int size_objective[2], 
-                  double* start_points, int size_start[2], 
+                  float* start_points, int size_start[2], 
                   int planner_type, int escalado_vectores);
 
 #endif // PLANNER_H
